+++ include/libIterativeRobot/commands/Base/BaseLinearMovement.h
@@ -12,13 +12,13 @@ class BaseLinearMovement : public libIterativeRobot::Command {
     void end();
     void interrupted();
     void blocked();
-    BaseLinearMovement(int leftTarget, int rightTarget);
-    BaseLinearMovement(int leftTarget, int rightTarget, double maxVel, double maxAccel);
-    BaseLinearMovement(int leftTarget, int rightTarget, bool absolute);
-    BaseLinearMovement(int leftTarget, int rightTarget, double maxVel, double maxAccel, bool absolute);
+    BaseLinearMovement(double leftTarget, double rightTarget);
+    BaseLinearMovement(double leftTarget, double rightTarget, double maxVel, double maxAccel);
+    BaseLinearMovement(double leftTarget, double rightTarget, bool absolute);
+    BaseLinearMovement(double leftTarget, double rightTarget, double maxVel, double maxAccel, bool absolute);
   private:
-    int leftTarget;
-    int rightTarget;
+    double leftTarget;
+    double rightTarget;
     double maxVel;
     double maxAccel;
     bool absolute;
